requirements.txt
streamlit==1.37.0
pandas==2.2.2
numpy==2.3.4
openpyxl==3.1.5
requests==2.32.5
openai==0.28
google-generativeai==0.8.5
streamlit-mermaid==0.3.0
langchain_community==0.4.1
langchain==1.0.2
langchain-text-splitters==1.0.0

# FastAPI dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
aiofiles==23.2.1
pydantic==2.6.1
pydantic-settings==2.1.0
